style: clear the clippy backlog, and make it a house rule
cargo build was warning-free but clippy had never been part of the gate, so about forty lints had accumulated across twenty-four files: needless borrows, redundant closures, &PathBuf where &Path does, useless vec! and format!, manual Option::map, map_or simplifications, missing Default impls on LinkQuality and RespawnPolicy, and redundant same-type casts. Type aliases stand in where type_complexity fired. None of it was a bug.

AGENTS.md now requires 'cargo clippy --all-targets -- -D warnings' to be clean, with targeted #[allow] plus a reason where a lint is genuinely wrong for the code -- never a blanket allow. Keeping it at zero is far cheaper than the next cleanup.

+++ examples/probe_mouse_tags.rs
@@ -64,7 +64,7 @@ fn keys_from(ranges: &[std::ops::RangeInclusive<u16>]) -> AttributeSet<KeyCode> 
 fn main() -> anyhow::Result<()> {
     // (a) today's set: every BTN_* except BTN_TOOL_*.
     let mut current = AttributeSet::<KeyCode>::new();
-    for code in 1..libc::KEY_MAX as u16 {
+    for code in 1..libc::KEY_MAX {
         let key = KeyCode::new(code);
         let name = format!("{:?}", key);
         if name.starts_with("BTN_") && !name.starts_with("BTN_TOOL_") {
